+++ packages/web/components/transactions/use-recent-transfers.ts
@@ -7,7 +7,7 @@ import {
 import { ChainIdHelper, isNumeric } from "@osmosis-labs/utils";
 
 import { ChainList } from "~/config/generated/chain-list";
-import { GetTransferStatusParams } from "~/integrations/bridges/interface";
+import { GetTransferStatusParams } from "~/integrations/bridges";
 import { useStore } from "~/stores";
 
 type RecentTransfer = {
