+++ packages/web/integrations/axelar/index.ts
@@ -1,1 +1,2 @@
+export * from "./types";
 export * from "./utils";
