+++ packages/web/integrations/bridges/chain.ts
@@ -1,0 +1,22 @@
+/** Human-displayable global source chain identifiers.
+ *  TODO: use global chain IDs instead of display names as keys
+ */
+export type SourceChain =
+  | "Bitcoin"
+  | "Bitcoin Testnet"
+  | "Aurora Testnet"
+  | "Avalanche"
+  | "Avalanche Fuji Testnet"
+  | "Binance Smart Chain"
+  | "BSC Testnet"
+  | "Ethereum"
+  | "Goerli Testnet"
+  | "Fantom"
+  | "Fantom Testnet"
+  | "Moonbeam"
+  | "Moonbase Alpha"
+  | "Polygon"
+  | "Mumbai"
+  | "Filecoin"
+  | "Filecoin Hyperspace"
+  | "Arbitrum";
