+++ packages/web/integrations/index.ts
@@ -1,3 +1,4 @@
-export * from "./bridge-info";
+export * from "./bridge";
+export * from "./fiat-ramps";
 export * from "./use-client-tx-event-toasts";
 export * from "./wallets";
