+++ packages/web/modals/select-asset-source.tsx
@@ -8,15 +8,15 @@ import {
   FiatRampDisplayInfos,
   FiatRampKey,
   ObservableWallet,
-  SourceChainKey,
 } from "~/integrations";
+import { SourceChain } from "~/integrations/bridges";
 import { ModalBase, ModalBaseProps } from "~/modals/base";
 
 /** Prompts user to connect from a list of wallets. Will onboard a user for an uninstalled wallet if the functionality is available. */
 export const SelectAssetSourceModal: FunctionComponent<
   ModalBaseProps & {
     initiallySelectedWalletId?: string;
-    desiredSourceKey?: SourceChainKey;
+    desiredSourceKey?: SourceChain;
     isWithdraw: boolean;
     wallets: ObservableWallet[];
     fiatRamps?: FiatRampKey[];
